remote-server / fixes

+++ ssh-proxy/remote-server/ssh-shared/entrypoint.sh
@@ -1,8 +1,8 @@
 #!/bin/sh
 
 # sync users
-cp -f /root/external/passwd /etc/passwd
-cp -f /root/external/shadow /etc/shadow
+[ -f /root/external/passwd ] && cp -f /root/external/passwd /etc/passwd
+[ -f /root/external/shadow ] && cp -f /root/external/shadow /etc/shadow
 
 # run server
 exec /usr/sbin/sshd -D "$@"
